extension: Move caching entirely into background
IndexedDB is not shared between different tabs, even within the same
extension. So it doesn't make sense for us to cache images from the Sonlet
tab and pass handles to the FB side for reading since the fb side won't
have access to the relevant IndexedDB.

Fix this by moving all of the caching into the background script. Sonlet
will now make a request to the background script to cache the images, and
the FB side will request the cached data from the background script (which
is returned in a Uint8Array), rather than trying to read it out of the
IndexedDB directly itself.