extension: background: Add script execution channel and setCaption action
We sometimes have the need to run in the execution context of the tab,
rather than in our extension sandbox. For example, for calling React
functions defined by the site (not the extension).

Add a new connect/port handler for script execution, along with the first
script execution action: "setCaption".